Add loadserver.sh
A little bit of TPS magic! This allows multiple maps to run independently, plus allows them to be shut down independently. It even allows a multi-user environment so various users have their own minetestserver running.
This was built for The Pixel Shadow's game servers since the MT community was missing a good gameloader.